move add_node_name function into lib
